Adding protothread fiddle
This adds an initial cut of a protothread implementation.  Needs to be
factored into something a little more clean.  It does work and has a
small test.  Note that this does some things differently than standard
protothreads with several macros that build a lot of boilerplate for you.
You get a function to call that sets up the protothread to run with the
arguments you pass.